Day 07
Leverages Kotlin’s Comparable<> to handle sorting of hands and cards. The type of a hand is mapped as an enum, which each enum entry having a predicate to determine if the hand matches it.

Some useful scenarios with jokers so make sure you run through a few of them:
AKJJJ -> FourOfAKind
AAKJJ -> FourOfAKind
AAAKJ -> FourOfAKind
AKQJJ -> ThreeOfAKind
AAQQJ -> FullHouse
JJJJJ -> FiveOfAKind

Once hands are sorted, simply a case of taking the rank of each hand and multiplying the bid and tallying it up.